requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
GitPython>=3.1.0
PyGithub>=1.55
tenacity>=8.0.0
